+++ src/app/api/auth/session/route.ts
@@ -1,7 +1,17 @@
 import { NextResponse } from "next/server"
 import { getServerSession } from "next-auth/next"
-import { authOptions } from "@/app/api/auth/[...nextauth]/route"
+import { authOptions } from "@/lib/auth"
 import prisma from "@/lib/prisma"
+
+export async function GET() {
+    try{
+        const session = await getServerSession(authOptions)
+        return NextResponse.json(session)
+    } catch (error) {
+        console.error("[SESSION_GET_ERROR]", error)
+        return new NextResponse("Internal Error", { status: 500 })
+    }
+}
 
 export async function DELETE() {
   try {
