+++ src/lib/auth.ts
@@ -1,11 +1,142 @@
 import { getServerSession } from "next-auth/next"
-import { authOptions } from "@/app/api/auth/[...nextauth]/route"
+import { NextAuthOptions } from "next-auth"
+import { PrismaAdapter } from "@next-auth/prisma-adapter"
+import GoogleProvider from "next-auth/providers/google"
+import CredentialsProvider from "next-auth/providers/credentials"
+import bcrypt from "bcryptjs"
+import prisma from "@/lib/prisma"
 
-export async function getSession() {
-  return await getServerSession(authOptions)
-}
+export const authOptions: NextAuthOptions = {
+    adapter: PrismaAdapter(prisma),
+    providers: [
+      GoogleProvider({
+        clientId: process.env.GOOGLE_CLIENT_ID!,
+        clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+        profile(profile) {
+          return {
+            id: profile.sub,
+            name: profile.name,
+            email: profile.email,
+            image: profile.picture,
+            emailVerified: profile.email_verified ? new Date() : null,
+          }
+        },
+      }),
+      CredentialsProvider({
+        name: "credentials",
+        credentials: {
+          email: { label: "Email", type: "text" },
+          password: { label: "Password", type: "password" },
+        },
+        async authorize(credentials) {
+          if (!credentials?.email || !credentials?.password) {
+            throw new Error("Invalid credentials")
+          }
+  
+          const user = await prisma.user.findUnique({
+            where: {
+              email: credentials.email,
+            },
+          })
+  
+          if (!user || !user.password) {
+            throw new Error("Invalid credentials")
+          }
+  
+          const isCorrectPassword = await bcrypt.compare(
+            credentials.password,
+            user.password
+          )
+  
+          if (!isCorrectPassword) {
+            throw new Error("Invalid credentials")
+          }
+  
+          await prisma.session.create({
+            data: {
+              userId: user.id,
+              expires: new Date(Date.now() + 30 * 24 * 60 * 60 * 1000),
+              sessionToken: crypto.randomUUID(),
+            },
+          })
+  
+          return user
+        },
+      }),
+    ],
+    session: {
+      strategy: "jwt",
+      maxAge: 30 * 24 * 60 * 60,
+    },
+    pages: {
+      signIn: "/login",
+    },
+    callbacks: {
+      async session({ session, token }) {
+        if (token && session.user) {
+          session.user.id = token.sub!
+          session.user.emailVerified = token.emailVerified as Date | null
+        }
+        return session
+      },
+      async jwt({ token, user, account }) {
+        if (user) {
+          token.sub = user.id
+          token.emailVerified = user.emailVerified
+        }
+        if (account) {
+          await prisma.account.upsert({
+            where: {
+              provider_providerAccountId: {
+                provider: account.provider,
+                providerAccountId: account.providerAccountId,
+              },
+            },
+            create: {
+              userId: user.id,
+              type: account.type,
+              provider: account.provider,
+              providerAccountId: account.providerAccountId,
+              access_token: account.access_token,
+              refresh_token: account.refresh_token,
+              expires_at: account.expires_at,
+              token_type: account.token_type,
+              scope: account.scope,
+              id_token: account.id_token,
+              session_state: account.session_state,
+            },
+            update: {
+              access_token: account.access_token,
+              refresh_token: account.refresh_token,
+              expires_at: account.expires_at,
+            },
+          })
+        }
+        return token
+      },
+    },
+    events: {
+      async signOut({ token }) {
+        try {
+          if (token?.sub) {
+            await prisma.session.deleteMany({
+              where: {
+                userId: token.sub,
+              },
+            })
+          }
+        } catch (error) {
+          console.error("Error deleting session:", error)
+        }
+      },
+    },
+  }
 
-export async function getCurrentUser() {
-  const session = await getSession()
-  return session?.user
-} 
+  export async function getSession() {
+    return await getServerSession(authOptions)
+  }
+  
+  export async function getCurrentUser() {
+    const session = await getSession()
+    return session?.user
+  }
