+++ src/types/next-auth.d.ts
@@ -1,9 +1,28 @@
-import { DefaultSession } from "next-auth"
+import "next-auth"
 
 declare module "next-auth" {
   interface Session {
     user: {
       id: string
-    } & DefaultSession["user"]
+      name: string | null
+      email: string
+      emailVerified: Date | null
+      image?: string | null
+    }
+  }
+
+  interface User {
+    id: string
+    name: string | null
+    email: string
+    emailVerified: Date | null
+    image?: string | null
+  }
+}
+
+declare module "next-auth/jwt" {
+  interface JWT {
+    id: string
+    emailVerified: Date | null
   }
 }
